[AMQCPP-30] updated output of priority in toString()
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@508908 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/connector/openwire/commands/Message.cpp
@@ -228,7 +228,7 @@ std::string Message::toString() const {
     stream << " Value of CorrelationId = " << this->getCorrelationId() << std::endl;
     stream << " Value of Persistent = " << this->isPersistent() << std::endl;
     stream << " Value of Expiration = " << this->getExpiration() << std::endl;
-    stream << " Value of Priority = " << this->getPriority() << std::endl;
+    stream << " Value of Priority = " << (int)this->getPriority() << std::endl;
     stream << " Value of ReplyTo is Below:" << std::endl;
     if( this->getReplyTo() != NULL ) {
         stream << this->getReplyTo()->toString() << std::endl;
